+++ docs/changes.md
@@ -3,6 +3,7 @@
 2024/03/16
   * 重构window manager将input device status对象放到基类。
   * 完善debugger_lldb修改请求超时时间。
+  * 完善debugger_lldb支持targetCreateCommands
 
 2024/03/15
   * 修复内存泄露。